requirements.txt
attrs==24.2.0
babel==2.16.0
certifi==2024.8.30
cffi==1.17.1
colorama==0.4.6
et_xmlfile==2.0.0
h11==0.14.0
idna==3.10
Jinja2==3.1.4
lxml==5.3.0
MarkupSafe==3.0.2
numpy==2.1.2
openpyxl==3.1.5
outcome==1.3.0.post0
pandas==2.2.3
prompt-toolkit==3.0.36
pycparser==2.22
Pygments==2.18.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
questionary==2.0.1
regex==2024.11.6
selenium==4.16.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
splinter==0.21.0
tkcalendar==1.6.1
tqdm==4.67.0
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
